[Add]: 게시글 작성 화면 추가
- WritePostView를 생성하여 게시글 작성 UI 구성
- 제목, 내용, 카테고리 선택이 가능한 폼 제공
- 등록 및 취소 동작 처리
- 입력 유무에 따라 경고(alert) 표시 및 등록 버튼 활성화 제어
- 등록 시 콜백을 통해 외부 ViewModel 또는 처리 로직과 연동 가능
